separate files, fix no shortcuts logic
