+++ src/shared/pages/welcomePage/style.css
@@ -16,10 +16,6 @@ span {
   font-weight: lighter;
 }
 
-[data-locale] {
-  transform: scaleY(0.925);
-}
-
 
 #mozilla-logo {
   margin-top: 12px;
@@ -29,7 +25,7 @@ span {
 .welcome-content {
   display: flex;
   padding: 60px;
-  align-items: start;
+  align-items: center;
 }
 
 #welcome-image {
@@ -46,6 +42,7 @@ span {
   min-width: 60%;
 }
 
+/* Mobile & Small Screen Styling */
 @media (max-width: 768px) {
   .welcome-content {
     flex-direction: column;
@@ -65,6 +62,7 @@ span {
   }
 }
 
+/* Title and Beta Styling */
 .welcome-header {
   display: flex;
   align-items: center;
@@ -78,10 +76,12 @@ span {
   margin-left: 10px;
 }
 
+/* Subtitle Styling */
 .welcome-info h2 {
   margin-top: 0px;
 }
 
+/* Browser List Styling (Firefox only) */
 .browser-list label {
   font-weight: 600;
   font-size: 0.7em;
@@ -96,6 +96,7 @@ span {
   margin-top: 5px;
 }
 
+/* Telemetry Styling (Both) and Always Private Styling (Chromium Only) */
 .telemetry,
 .always-private {
   display: flex;
@@ -109,6 +110,12 @@ span {
   margin-top: 2px;
 }
 
+/* Shortcuts List Styling */
+#shortcuts-list {
+  margin-top: -20px;
+}
+
+/* Slider Styling */
 .switch {
   position: relative;
   width: 40px;
@@ -160,10 +167,3 @@ input:checked + .slider:before {
   border-radius: 50%;
 }
 
-#shortcuts-list {
-  margin-top: -20px;
-}
-
-.shortcut-item {
-  margin-top: 0px;
-}
